file restructuring, new home file created
