+++ src/app/home.py
@@ -1,0 +1,2 @@
+# Handles home page functionality
+
