+++ src/auth/email_handler.py
